file restructuring, fixed deprecated code
