--- app/src/main/java/com/droidheat/amoledbackgrounds/adapters/ViewPagerAdapter.java
@@ -1,33 +1,0 @@
-package com.droidheat.amoledbackgrounds.adapters;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentPagerAdapter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ViewPagerAdapter extends FragmentPagerAdapter {
-	private final List<Fragment> mFragmentList = new ArrayList<>();
-	
-	public ViewPagerAdapter(FragmentManager manager) {
-		super(manager);
-	}
-	
-	@NonNull
-	@Override
-	public Fragment getItem(int position) {
-		return mFragmentList.get(position);
-	}
-	
-	@Override
-	public int getCount() {
-		return mFragmentList.size();
-	}
-	
-	public void addFragment(Fragment fragment) {
-		mFragmentList.add(fragment);
-	}
-	
-}
