--- app/src/main/res/drawable/design_black_bg_white_border.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <stroke
-        android:width="1dp"
-        android:color="#44ffffff"
-        />
-
-    <corners
-        android:radius="3dp" />
-
-    <solid
-        android:color="@color/colorBackground"/>
-
-</shape>
