+++ app/src/main/res/layout/fragment_wallpaper.xml
@@ -20,7 +20,6 @@
             android:textAlignment="center">
 
             <TextView
-                android:id="@+id/text43121"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_centerVertical="true"
