+++ app/src/main/res/layout/my_downloads_list_item.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="350dp"
     android:layout_margin="5dp">
@@ -9,7 +10,6 @@
         android:layout_width="match_parent"
         android:layout_height="350dp"
         android:background="@color/colorBackground">
-
 
 
         <TextView
@@ -26,10 +26,9 @@
             android:textStyle="bold"
             android:layout_marginEnd="6dp"
             android:maxLines="2"
-            android:text="asdfasdsadasdasdas sadasd asd asdasd asd asdas jdsfdasdasdas"
+            android:text="@string/app_name"
             android:ellipsize="marquee"
-            android:textColor="@android:color/white"
-            />
+            android:textColor="@android:color/white" />
 
         <ImageView
             android:layout_width="wrap_content"
@@ -41,7 +40,7 @@
             android:layout_alignParentEnd="true"
             android:layout_alignTop="@+id/title"
             android:background="?android:attr/selectableItemBackgroundBorderless"
-            />
+            android:contentDescription="@string/app_name" />
 
         <RelativeLayout
             android:layout_width="match_parent"
@@ -56,7 +55,7 @@
                 android:scaleType="centerCrop"
                 android:layout_margin="1dp"
                 android:id="@+id/preview"
-                />
+                android:contentDescription="@string/app_name" />
 
             <ImageView
                 android:layout_width="48dp"
@@ -66,8 +65,9 @@
                 android:layout_alignParentBottom="true"
                 android:id="@+id/download"
                 android:background="#44000000"
-                android:tint="@android:color/white"
-                android:padding="12dp"/>
+                android:padding="12dp"
+                app:tint="@android:color/white"
+                android:contentDescription="@string/app_name" />
 
             <ProgressBar
                 android:layout_width="48dp"
@@ -76,7 +76,7 @@
                 android:padding="12dp"
                 android:layout_alignParentEnd="true"
                 android:layout_alignParentBottom="true"
-                android:background="#44000000"/>
+                android:background="#44000000" />
 
         </RelativeLayout>
