+++ app/src/main/res/values/ids.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <item name="popup_share" type="id"></item>
-    <item name="popup_open_file_location" type="id"></item>
-    <item name="popup_delete" type="id"></item>
+    <item name="popup_share" type="id" />
+    <item name="popup_delete" type="id" />
 </resources>
