+++ app/src/main/res/xml/provider_paths.xml
@@ -1,4 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<paths xmlns:android="http://schemas.android.com/apk/res/android">
-    <external-path name="external_files" path="*"/>
+<paths>
+    <external-path
+        name="external_files"
+        path="*" />
 </paths>
